Use imported Cloudflare handler in env API
