+++ packages/astro/src/core/app/index.ts
@@ -1,3 +1,4 @@
+export type { RoutesList } from '../../types/astro.js';
 export { App } from './app.js';
 export { BaseApp, type RenderErrorOptions, type RenderOptions } from './base.js';
 export { deserializeManifest, fromRoutingStrategy, toRoutingStrategy } from './common.js';
