+++ packages/astro/src/vite-plugin-app/pipeline.ts
@@ -62,8 +62,7 @@ export class AstroServerPipeline extends Pipeline {
 			logger,
 			manifest,
 			settings,
-		}: Pick<AstroServerPipeline, 'loader' | 'logger' | 'manifest' | 'settings'>
-		
+		}: Pick<AstroServerPipeline, 'loader' | 'logger' | 'manifest' | 'settings'>,
 	) {
 		const pipeline = new AstroServerPipeline(loader, logger, manifest, settings);
 		pipeline.routesList = manifestData;
