+++ packages/integrations/cloudflare/src/entrypoints/server.ts
@@ -1,24 +1,25 @@
-import type { ExecutionContext, ExportedHandlerFetchHandler } from '@cloudflare/workers-types';
-import type { SSRManifest } from 'astro';
+/// <reference types="vite/client" />
 
-import { App } from 'astro/app';
-import { handle } from '../utils/handler.js';
+import type { ExportedHandler } from '@cloudflare/workers-types';
+import type { RouteInfo, SSRManifest } from 'astro';
+import { App, type BaseApp, createConsoleLogger, DevApp, type RoutesList } from 'astro/app';
+import { type Env, handle } from '../utils/handler.js';
 
-type Env = {
-	[key: string]: unknown;
-	ASSETS: { fetch: (req: Request | string) => Promise<Response> };
-};
+export function createExports(manifest: SSRManifest, routes: RouteInfo[]) {
+	let app: BaseApp;
 
-export function createExports(manifest: SSRManifest) {
-	const app = new App(manifest);
-
-	const fetch = async (
-		request: Parameters<ExportedHandlerFetchHandler>[0],
-		env: Env,
-		context: ExecutionContext,
-	) => {
-		return await handle(manifest, app, request, env, context);
+	if (import.meta.env.DEV) {
+		const routesList: RoutesList = { routes: routes.map((r: RouteInfo) => r.routeData) };
+		const logger = createConsoleLogger('debug');
+		app = new DevApp(manifest, true, logger, routesList);
+	} else {
+		app = new App(manifest);
+	}
+	return {
+		default: {
+			async fetch(request, env, ctx) {
+				return await handle(manifest, app, request, env, ctx);
+			},
+		} satisfies ExportedHandler<Env>,
 	};
-
-	return { default: { fetch } };
 }
