+++ packages/integrations/cloudflare/test/fixtures/vite-plugin/astro.config.mjs
@@ -5,8 +5,5 @@ import { defineConfig } from 'astro/config';
 
 export default defineConfig({
 	adapter: cloudflare({
-		workerEntryPoint: {
-			path: 'src/worker.ts',
-		}
 	}),
 });
