+++ packages/integrations/cloudflare/test/fixtures/vite-plugin/src/dev.ts
@@ -1,5 +1,5 @@
 import { manifest } from "astro:serialized-manifest";
 import { routes } from "astro:routes"
-import { createExports } from "./worker";
+import { createExports } from "@astrojs/cloudflare/entrypoints/server.js";
 
 export default createExports(manifest, routes).default
